backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Stanley Hurst <sdhurst@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "^0.100"
pydantic = { version = "^2.0", extras = ["email"] }
uvicorn = "^0.22"
fastapi-utils = "^0.8.0"
pdfplumber = "*"
python-docx = "*"
pillow = "*"

asyncpg = "*"
psycopg2-binary = "*"

typing-inspect = "^0.9.0"
pdf2image = "^1.17.0"
opencv-python-headless = "^4.11.0.86"
numpy = "^2.3.1"
pytesseract = "^0.3.13"
email-validator = "^2.2.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
python-dotenv = "^1.1.1"
python-multipart = "*"
python-jose = "^3.3.0"
[tool.poetry.group.dev.dependencies]
pytest = "*"
black = "*"
isort = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

